[2023-03-22] 1. 가게 평균 평점 추가 (jpql 사용)
2. cors 세팅

+++ src/main/java/com/example/jmcbackend/review/dto/ReviewDto.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Builder
 public class ReviewDto {
 
-    private Byte reviewScore;
+    private Float reviewScore;
     private String reviewText;
 
 }
